@@ -89,36 +89,37 @@
     "commander": "2.20.0",
     "electron-window-state": "5.0.3",
     "post-robot": "https://github.com/dragonbane0/post-robot/tarball/dad59c1c0f6266d0dd4bd35e0fb7137d45c6c773",
-    "tree-kill": "1.2.1"
+    "tree-kill": "1.2.1",
+    "tslib": "^1.10.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.29",
-    "@angular/animations": "8.2.14",
-    "@angular/cdk": "8.2.3",
-    "@angular/cli": "8.3.29",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/flex-layout": "8.0.0-beta.27",
-    "@angular/forms": "8.2.14",
-    "@angular/language-service": "8.2.14",
-    "@angular/material": "8.2.3",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular/animations": "9.1.13",
+    "@angular/cdk": "9.1.3",
+    "@angular/cli": "9.1.15",
+    "@angular/common": "9.1.13",
+    "@angular/compiler": "9.1.13",
+    "@angular/compiler-cli": "9.1.13",
+    "@angular/core": "9.1.13",
+    "@angular/flex-layout": "9.0.0-beta.31",
+    "@angular/forms": "9.1.13",
+    "@angular/language-service": "9.1.13",
+    "@angular/material": "9.1.3",
+    "@angular/platform-browser": "9.1.13",
+    "@angular/platform-browser-dynamic": "9.1.13",
+    "@angular/router": "9.1.13",
     "@angularclass/hmr": "2.1.3",
     "@fortawesome/fontawesome-free": "5.9.0",
-    "@nebular/eva-icons": "4.6.0",
-    "@nebular/theme": "4.6.0",
+    "@nebular/eva-icons": "5.1.0",
+    "@nebular/theme": "5.1.0",
     "@types/electron": "1.6.10",
     "@types/jasmine": "2.8.16",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "8.9.5",
+    "@types/node": "^12.11.1",
     "angular-dual-listbox": "4.7.0",
-    "angular-file": "1.3.2",
+    "angular-file": "3.1.2",
     "browserslist": "^4.21.4",
-    "codelyzer": "^5.0.1",
+    "codelyzer": "^5.1.2",
     "core-js": "3.25.1",
     "crc": "3.8.0",
     "electron": "4.2.6",
@@ -146,12 +147,11 @@
     "socicon": "3.0.5",
     "tcp-ping": "0.1.1",
     "ts-node": "7.0.1",
-    "tslib": "1.10.0",
     "tslint": "5.11.0",
     "typeface-exo": "0.0.22",
-    "typescript": "3.5.3",
+    "typescript": "3.8.3",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "overrides": {
     "browserslist": {
